requirements.txt
asyncio
asyncpg
audioop-lts
attrs
blinker
click
deep-translator
discord.py
Flask
frozenlist
fuzzywuzzy
PyNaCl
pytz
requests
sprite
typing_extensions
urllib3
yarl
zipp
pyfiglet
python-dateutil